실습용/06_vectordb/requirements.txt
pymupdf
pillow
requests
python-dotenv
python-docx
pandas
openpyxl
